Fix missing identity popup integration for existing private windows.
Due to bug 1196577 the browserWindows object of sdk/windows does not
contain the private windows that are opened before the add-on has
loaded. This is especially problematic in always-on private browsing
mode.

Luckily the sdk/window/utils module contains a method that allows
add-ons to list all windows, including existing private ones. This
changeset coverts the add-on to use that instead of browserWindows
object.